+++ src/applications/item-pile-inventory-app/CurrencyList.svelte
@@ -11,13 +11,13 @@
   const editQuantities = store.editQuantities;
 
   async function addCurrency() {
-    const result = await DropCurrencyDialog.show(store.recipient, store.source);
+    const result = await DropCurrencyDialog.show(store.recipient, store.actor);
     if (!result) return;
     if (!foundry.utils.isObjectEmpty(result.attributes)) {
-      await game.itempiles.transferAttributes(store.recipient, store.source, result.attributes, { interactionId: store.interactionId })
+      await game.itempiles.transferAttributes(store.recipient, store.actor, result.attributes, { interactionId: store.interactionId })
     }
     if (result.items.length) {
-      await game.itempiles.transferItems(store.recipient, store.source, result.items, { interactionId: store.interactionId })
+      await game.itempiles.transferItems(store.recipient, store.actor, result.items, { interactionId: store.interactionId })
     }
   }