+++ src/systems/forbidden-lands.js
@@ -1,21 +1,21 @@
 export default {
   
-  "VERSION": "1.0.1",
+  "VERSION": "1.0.0",
   
   // The actor class type is the type of actor that will be used for the default item pile actor that is created on first item drop.
   "ACTOR_CLASS_TYPE": "character",
   
   // The item quantity attribute is the path to the attribute on items that denote how many of that item that exists
-  "ITEM_QUANTITY_ATTRIBUTE": "data.quantity.value",
+  "ITEM_QUANTITY_ATTRIBUTE": "data.quantity",
   
   // The item price attribute is the path to the attribute on each item that determine how much it costs
-  "ITEM_PRICE_ATTRIBUTE": "data.price.gc",
+  "ITEM_PRICE_ATTRIBUTE": "data.cost",
   
   // Item types and the filters actively remove items from the item pile inventory UI that users cannot loot, such as spells, feats, and classes
   "ITEM_FILTERS": [
     {
       "path": "type",
-      "filters": "career,container,critical,disease,injury,mutation,prayer,psychology,talent,skill,spell,trait,extendedTest,vehicleMod,cargo"
+      "filters": "criticalInjury,monsterAttack,monsterTalent"
     }
   ],
   
@@ -27,36 +27,37 @@ export default {
   // In the case of items, it is recommended you export the item with `.toObject()` and strip out any module data
   "CURRENCIES": [
     {
-      type: "item",
-      name: "Gold Crown",
+      type: "attribute",
+      name: "Gold",
       img: "icons/commodities/currency/coin-embossed-crown-gold.webp",
       abbreviation: "{#}GP",
       data: {
-        uuid: "Compendium.wfrp4e.basic.UHArNq3Vuu7Mj5AB"
+        path: "data.currency.gold",
       },
       primary: true,
       exchangeRate: 1
     },
     {
-      type: "item",
-      name: "Silver Pieces",
+      type: "attribute",
+      name: "Silver",
       img: "icons/commodities/currency/coin-engraved-moon-silver.webp",
       abbreviation: "{#}SP",
       data: {
-        uuid: "Compendium.wfrp4e.basic.AtRuawWCHjgQxzdn"
+        path: "data.currency.silver",
       },
       primary: false,
       exchangeRate: 0.1
     },
     {
-      type: "item",
-      name: "Brass Penny",
+      type: "attribute",
+      name: "Copper",
       img: "icons/commodities/currency/coin-engraved-waves-copper.webp",
       abbreviation: "{#}CP",
       data: {
-        uuid: "Compendium.wfrp4e.basic.1BdNwUF0SfiCotWB"
+        path: "data.currency.copper",
       },
       primary: false,
       exchangeRate: 0.01
-    }]
+    }
+  ]
 }