+++ src/systems/pf1.js
@@ -1,12 +1,15 @@
 export default {
   
-  "VERSION": "1.0.1",
+  "VERSION": "1.0.2",
   
   // The actor class type is the type of actor that will be used for the default item pile actor that is created on first item drop.
   "ACTOR_CLASS_TYPE": "npc",
   
   // The item quantity attribute is the path to the attribute on items that denote how many of that item that exists
   "ITEM_QUANTITY_ATTRIBUTE": "data.quantity",
+  
+  // The item price attribute is the path to the attribute on each item that determine how much it costs
+  "ITEM_PRICE_ATTRIBUTE": "data.price",
   
   // Item types and the filters actively remove items from the item pile inventory UI that users cannot loot, such as spells, feats, and classes
   "ITEM_FILTERS": [