+++ src/systems/swse.js
@@ -1,6 +1,6 @@
 export default {
   
-  "VERSION": "1.0.1",
+  "VERSION": "1.0.0",
   
   // The actor class type is the type of actor that will be used for the default item pile actor that is created on first item drop.
   "ACTOR_CLASS_TYPE": "character",
@@ -8,31 +8,25 @@ export default {
   // The item quantity attribute is the path to the attribute on items that denote how many of that item that exists
   "ITEM_QUANTITY_ATTRIBUTE": "data.quantity",
   
-  // The item price attribute is the path to the attribute on each item that determine how much it costs
-  "ITEM_PRICE_ATTRIBUTE": "data.price",
-  
   // Item types and the filters actively remove items from the item pile inventory UI that users cannot loot, such as spells, feats, and classes
   "ITEM_FILTERS": [
     {
       "path": "type",
-      "filters": "edge,hindrance,skill,power,ability"
+      "filters": "affiliation,background,class,beast quality,destiny,feat,forcePower,forceRegimen,forceSecret,forceTechnique,species,talent,template,trait,vehicleSystem"
     }
   ],
   
   // Item similarities determines how item piles detect similarities and differences in the system
-  "ITEM_SIMILARITIES": ["name", "type"],
+  "ITEM_SIMILARITIES": ["type", "name", "strippable", "hasPrerequisites", "modifiable", "hasLevels"],
   
-  // Currencies in item piles is a versatile system that can accept actor attributes (a number field on the actor's sheet) or items (actual items in their inventory)
-  // In the case of attributes, the path is relative to the "actor.data"
-  // In the case of items, it is recommended you export the item with `.toObject()` and strip out any module data
   "CURRENCIES": [
     {
       type: "attribute",
-      name: "SWADE.Currency",
+      name: "Credits",
       img: "icons/svg/coins.svg",
-      abbreviation: "{#}T",
+      abbreviation: "{#}C",
       data: {
-        path: "data.details.currency",
+        path: "data.common.credits",
       },
       primary: true,
       exchangeRate: 1